requirements.txt
acres
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
anyio
apispec
asn1crypto
async-timeout
attrs
bech32
blinker
boto3
botocore
cachetools
certifi
cffi
charset-normalizer
ci-info
click
cmdstanpy
colorama
configobj
configparser
contourpy
cosmpy
cryptography
cycler
distlib
distro
ecdsa
etelemetry
exceptiongroup
filelock
fitz
Flask
fonttools
frozenlist
gitdb
GitPython
googleapis-common-protos
together
grpcio
h11
holidays
httpcore
httplib2
httpx
idna
importlib_metadata
importlib_resources
isodate
itsdangerous
Jinja2
jiter
jmespath
jsonschema
jsonschema-specifications
kiwisolver
looseversion
lxml
markdown-it-py
MarkupSafe
matplotlib
mdurl
msgpack
multidict
narwhals
networkx
nibabel
nipype
numpy
# openai  # Replaced with together.ai
packaging
pandas
pathlib
pdf2image
pillow
pkgutil_resolve_name
platformdirs
plotly
propcache
prophet
protobuf
prov
puremagic
pyarrow
pycparser
pycryptodome
pydantic
pydantic-settings
pydantic_core
pydeck
pydot
Pygments
PyJWT
pyOpenSSL
pyparsing
PyPDF2
pytesseract
python-dateutil
python-dotenv
python-multipart
pytz
pyxnat
rdflib
referencing
requests
rich
rpds-py
s3transfer
scipy
simplejson
six
smmap
sniffio
snowflake-connector-python
sortedcontainers
stanio
streamlit
tenacity
toml
tomlkit
tornado
tqdm
traits
typing-inspection
typing_extensions
tzdata
uagents
urllib3
uvicorn
virtualenv
watchdog
websockets
Werkzeug
yarl
zipp
